P2.6: drop @deprecated AuditService delegators, move @internal helpers to AuditRecorder
