+++ src/services/AuditService.php
@@ -10,34 +10,14 @@
 
 namespace superbig\audit\services;
 
-use Craft;
 use craft\base\Component;
 use craft\base\ElementInterface;
-use craft\base\Plugin;
-use craft\base\PluginInterface;
-use craft\elements\User;
-use craft\events\BackupEvent;
-use craft\events\ConfigEvent;
-use craft\events\EntryTypeEvent;
-use craft\events\FieldEvent;
-use craft\events\RestoreEvent;
-use craft\events\RouteEvent;
-use craft\events\SectionEvent;
-use craft\events\UserAssignGroupEvent;
-use craft\events\UserEvent;
-use craft\events\UserGroupEvent;
-use craft\events\UserGroupPermissionsEvent;
-use craft\events\UserPermissionsEvent;
-use craft\helpers\Json;
 use craft\helpers\Template;
-use craft\queue\jobs\ResaveElements;
 
 use DateTime;
 use superbig\audit\Audit;
-use superbig\audit\events\SnapshotEvent;
 use superbig\audit\models\AuditModel;
 use superbig\audit\records\AuditRecord;
-use yii\base\Exception;
 
 /**
  * @author    Superbig
@@ -74,7 +54,6 @@ class AuditService extends Component
      */
     public function getEventById(int $id): ?AuditModel
     {
-        $models = null;
         $record = AuditRecord::findOne($id);
 
         if (!$record) {
@@ -129,160 +108,6 @@ class AuditService extends Component
                           ->count();
     }
 
-    /**
-     * @param ElementInterface $element
-     * @param bool $isNew
-     *
-     * @return bool
-     * @deprecated Use Audit::$plugin->elementHandler->onSaveElement() instead.
-     */
-    public function onSaveElement(ElementInterface $element, bool $isNew = false): bool
-    {
-        return Audit::$plugin->elementHandler->onSaveElement($element, $isNew);
-    }
-
-    /**
-     * @param ElementInterface $element
-     *
-     * @return bool
-     * @deprecated Use Audit::$plugin->elementHandler->onDeleteElement() instead.
-     */
-    public function onDeleteElement(ElementInterface $element): bool
-    {
-        return Audit::$plugin->elementHandler->onDeleteElement($element);
-    }
-
-    /**
-     * @return bool
-     * @deprecated Use Audit::$plugin->userHandler->onLogin() instead.
-     */
-    public function onLogin(): bool
-    {
-        return Audit::$plugin->userHandler->onLogin();
-    }
-
-    /**
-     * @return bool
-     * @deprecated Use Audit::$plugin->userHandler->onBeforeLogout() instead.
-     */
-    public function onBeforeLogout(): bool
-    {
-        return Audit::$plugin->userHandler->onBeforeLogout();
-    }
-
-    /**
-     * @param string          $event
-     * @param PluginInterface $plugin
-     *
-     * @return bool
-     */
-    /**
-     * @deprecated Use Audit::$plugin->pluginHandler->onPluginEvent() instead
-     */
-    public function onPluginEvent(string $event, PluginInterface $plugin): bool
-    {
-        return Audit::$plugin->pluginHandler->onPluginEvent($event, $plugin);
-    }
-
-    /**
-     * @param AuditModel $auditModel
-     * @param            $snapshot
-     *
-     * @return array
-     * @internal Public for use by handler services in services/handlers/.
-     */
-    public function afterSnapshot(AuditModel $auditModel, $snapshot): array
-    {
-        $event = new SnapshotEvent([
-            'audit' => $auditModel,
-            'snapshot' => $snapshot,
-        ]);
-
-        $this->trigger(self::EVENT_SNAPSHOT, $event);
-
-        return $event->snapshot;
-    }
-
-    /**
-     * @return AuditModel
-     * @internal Public for use by handler services in services/handlers/. Do not call from outside the plugin.
-     */
-    public function _getStandardModel(): AuditModel
-    {
-        $app = Craft::$app;
-        $request = $app->getRequest();
-        $model = new AuditModel();
-        $model->siteId = $app->getSites()->currentSite->id;
-
-        if (!$request->isConsoleRequest) {
-            $session = $app->getSession();
-            $model->sessionId = $session->getId();
-            $model->ip = $request->getUserIP();
-            $model->userAgent = $request->getUserAgent();
-
-            if ($identity = $app->getUser()->getIdentity()) {
-                $model->userId = $identity->id;
-
-                $model->snapshot = [
-                    'userId' => $model->userId,
-                ];
-            }
-        }
-
-        return $model;
-    }
-
-    /**
-     * @param AuditModel $model
-     * @param bool $unique
-     *
-     * @return bool
-     */
-    public function _saveRecord(AuditModel &$model, bool $unique = true): bool
-    {
-        try {
-            if ($model->id) {
-                $record = AuditRecord::findOne($model->id);
-            } else {
-                $record = new AuditRecord();
-            }
-
-            $record->event = $model->event;
-            $record->title = $model->title;
-            $record->parentId = $model->parentId;
-            $record->userId = $model->userId;
-            $record->elementId = $model->elementId;
-            $record->elementType = $model->elementType;
-            $record->ip = $model->ip;
-            $record->userAgent = $model->userAgent;
-            $record->siteId = $model->siteId;
-            $record->snapshot = Json::encode($model->snapshot ?: []);
-            $record->sessionId = $model->sessionId;
-
-            if (!$record->save()) {
-                Craft::error(
-                    Craft::t('audit', 'An error occured when saving audit log record: {error}',
-                        [
-                            'error' => print_r($record->getErrors(), true),
-                        ]),
-                    'audit');
-            }
-
-            $model->id = $record->id;
-
-            return true;
-        } catch (Exception $e) {
-            Craft::error(
-                Craft::t('audit', 'An error occured when saving audit log record: {error}',
-                    [
-                        'error' => $e->getMessage(),
-                    ]),
-                'audit');
-
-            return false;
-        }
-    }
-
     public function outputObjectAsTable($input, $end = true): string|\Twig\Markup
     {
         $output = '<table class="audit-snapshot-table">';
@@ -322,257 +147,5 @@ class AuditService extends Component
         AuditRecord::deleteAll(['<=', 'dateCreated', $date]);
 
         return $count;
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->elementHandler->onBeforeResave() instead.
-     */
-    public function onBeforeResave(ResaveElements $job): bool
-    {
-        return Audit::$plugin->elementHandler->onBeforeResave($job);
-    }
-
-    /**
-     * @param string $elementType
-     *
-     * @return mixed
-     * @deprecated Use Audit::$plugin->elementHandler->getParentId() instead.
-     */
-    public function getParentId(string $elementType = ''): mixed
-    {
-        return Audit::$plugin->elementHandler->getParentId($elementType);
-    }
-
-    /**
-     * @param ResaveElements $job
-     *
-     * @return mixed
-     * @deprecated Use Audit::$plugin->elementHandler->onResaveEnd() instead.
-     */
-    public function onResaveEnd(ResaveElements $job): mixed
-    {
-        return Audit::$plugin->elementHandler->onResaveEnd($job);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->elementHandler->getParentIdKey() instead.
-     */
-    public function getParentIdKey($elementType = ''): string
-    {
-        return Audit::$plugin->elementHandler->getParentIdKey($elementType);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->routeHandler->onSaveRoute() instead
-     */
-    public function onSaveRoute(RouteEvent $event)
-    {
-        return Audit::$plugin->routeHandler->onSaveRoute($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->routeHandler->onDeleteRoute() instead
-     */
-    public function onDeleteRoute(RouteEvent $event)
-    {
-        return Audit::$plugin->routeHandler->onDeleteRoute($event);
-    }
-
-    /**
-     * @internal Public for use by handler services in services/handlers/.
-     */
-    public function catchSaveError(callable $callable)
-    {
-        try {
-            return $callable();
-        } catch (\Exception $e) {
-            $this->logSaveError($e);
-
-            return false;
-        }
-    }
-
-    private function logSaveError(\Exception $e): void
-    {
-        Craft::error(
-            Craft::t('audit', 'Error when logging: {error}', ['error' => $e->getMessage()]),
-            __METHOD__
-        );
-    }
-
-    // =========================================================================
-    // User Security Events
-    // =========================================================================
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserActivated() instead.
-     */
-    public function onUserActivated(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserActivated($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserDeactivated() instead.
-     */
-    public function onUserDeactivated(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserDeactivated($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserSuspended() instead.
-     */
-    public function onUserSuspended(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserSuspended($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserUnsuspended() instead.
-     */
-    public function onUserUnsuspended(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserUnsuspended($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserLocked() instead.
-     */
-    public function onUserLocked(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserLocked($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userHandler->onUserUnlocked() instead.
-     */
-    public function onUserUnlocked(UserEvent $event): bool
-    {
-        return Audit::$plugin->userHandler->onUserUnlocked($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userGroupHandler->onUserGroupsAssigned() instead.
-     */
-    public function onUserGroupsAssigned(UserAssignGroupEvent $event): bool
-    {
-        return Audit::$plugin->userGroupHandler->onUserGroupsAssigned($event);
-    }
-
-    // =========================================================================
-    // Permission Events
-    // =========================================================================
-
-    /**
-     * @deprecated Use Audit::$plugin->userGroupHandler->onUserPermissionsSaved() instead.
-     */
-    public function onUserPermissionsSaved(UserPermissionsEvent $event): bool
-    {
-        return Audit::$plugin->userGroupHandler->onUserPermissionsSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userGroupHandler->onGroupPermissionsSaved() instead.
-     */
-    public function onGroupPermissionsSaved(UserGroupPermissionsEvent $event): bool
-    {
-        return Audit::$plugin->userGroupHandler->onGroupPermissionsSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userGroupHandler->onUserGroupSaved() instead.
-     */
-    public function onUserGroupSaved(UserGroupEvent $event): bool
-    {
-        return Audit::$plugin->userGroupHandler->onUserGroupSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->userGroupHandler->onUserGroupDeleted() instead.
-     */
-    public function onUserGroupDeleted(UserGroupEvent $event): bool
-    {
-        return Audit::$plugin->userGroupHandler->onUserGroupDeleted($event);
-    }
-
-    // =========================================================================
-    // Schema Events
-    // =========================================================================
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onFieldSaved() instead
-     */
-    public function onFieldSaved(FieldEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onFieldSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onFieldDeleted() instead
-     */
-    public function onFieldDeleted(FieldEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onFieldDeleted($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onSectionSaved() instead
-     */
-    public function onSectionSaved(SectionEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onSectionSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onSectionDeleted() instead
-     */
-    public function onSectionDeleted(SectionEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onSectionDeleted($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onEntryTypeSaved() instead
-     */
-    public function onEntryTypeSaved(EntryTypeEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onEntryTypeSaved($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->schemaHandler->onEntryTypeDeleted() instead
-     */
-    public function onEntryTypeDeleted(EntryTypeEvent $event): bool
-    {
-        return Audit::$plugin->schemaHandler->onEntryTypeDeleted($event);
-    }
-
-    // =========================================================================
-    // Database Events
-    // =========================================================================
-
-    /**
-     * @deprecated Use Audit::$plugin->backupHandler->onBackupCreated() instead
-     */
-    public function onBackupCreated(BackupEvent $event): bool
-    {
-        return Audit::$plugin->backupHandler->onBackupCreated($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->backupHandler->onBackupRestored() instead
-     */
-    public function onBackupRestored(RestoreEvent $event): bool
-    {
-        return Audit::$plugin->backupHandler->onBackupRestored($event);
-    }
-
-    /**
-     * @deprecated Use Audit::$plugin->settingsHandler->onSettingsChanged() instead
-     */
-    public function onSettingsChanged(ConfigEvent $event, string $settingsType): bool
-    {
-        return Audit::$plugin->settingsHandler->onSettingsChanged($event, $settingsType);
     }
 }
